chore (homeManager arch): module added

+++ hosts/homeManager/x86_64-linux/home.nix
@@ -19,7 +19,7 @@
     myBackupScript = lib.hm.dag.entryAfter ["writeBoundary"] ''
       # Custom backup logic if needed
       for f in ${config.home.homeDirectory}/.bashrc ${config.home.homeDirectory}/.zshrc ${config.home.homeDirectory}/.config/systemd/user/tmux.service; do
-        if [ -f "$f" ]; then
+        if [[ -f "$f" && -L "$f" ]]; then
           mv "$f" "$f.hm-bak"
         fi
       done
